--- src/app/home/page.tsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <h1>Home</h1>;
-}
