+++ src/pages/api/signIn/index.ts
@@ -2,20 +2,19 @@ import { Amplify, Auth } from 'aws-amplify';
 import { amplifyConfig } from 'config/cognito';
 import { NextApiResponse } from 'next';
 import { BaseResponse } from 'services/httpClient/types';
+import { SignInApiRequest } from 'shared/types/api/signIn';
 import { Logger } from 'utils';
-
-import { SignInApiRequest } from './types';
 
 Amplify.configure(amplifyConfig);
 
-export default async function handler<T>(
+export default async function handler(
   req: SignInApiRequest,
-  res: NextApiResponse<BaseResponse<T>>
+  res: NextApiResponse<BaseResponse<void>>
 ) {
   if (req.method === 'POST') {
     const { username, password } = req.body;
     try {
-      const response = (await Auth.signIn({ username, password })) as T;
+      const response = await Auth.signIn({ username, password });
 
       return res.status(200).json({ data: response, ok: true, status: 200 });
     } catch (error) {
